Re-organize mapping util functions
